Add inactive_contributors task to prune inactive contributors
This should be run periodically to identify inactive contributors. We
plan to remove privileged access for inactive contributors to reduce the
risk of malicious changes.

Queries the [githubarchive](https://www.gharchive.org) dataset via
BigQuery to identify users that have had any activity on repositories in
the configured org in the last 180 days.

Signed-off-by: Michael Smith <michael.smith@puppet.com>