More informative reporting
This now includes a list of any offenses in the status report. It will
flag on teams that don't exist in our org and rules that don't match any
teams. It will also list users assigned in the CODEOWNERS files so that
people can keep track of users who've moved on.